+++ RestServerFecurity/models/server.js
@@ -9,11 +9,11 @@ class Server {
         this.app  = express();
         this.port = process.env.PORT;
 
-        this.usuariosPath = '/api/users';
+        this.userPath = '/api/users';
 
 
         // Conectar a base de datos
-        this.conectarDB();
+        this.connectDB();
 
         // Middlewares
         this.middlewares();
@@ -22,7 +22,7 @@ class Server {
         this.routes();
     }
 
-    async conectarDB() {
+    async connectDB() {
         await dbConnectionFec();
     }
 
@@ -42,12 +42,12 @@ class Server {
 
     routes() {
         
-        this.app.use( this.usuariosPath, require('../routes/usuarios'));
+        this.app.use( this.userPath, require('../routes/users_routes'));
     }
 
     listen() {
         this.app.listen( this.port, () => {
-            console.log('Servidor corriendo en puerto', this.port );
+            console.log('Server running in port', this.port );
         });
     }
